--- src/utils/CarNameParse.js
@@ -1,9 +1,0 @@
-function CarNameParse(str) {
-  let carNames = str.split(',');
-  carNames = carNames.map((name) => {
-    return name.trim();
-  });
-  return carNames.slice();
-}
-
-module.exports = CarNameParse;
